main/requirements.txt
langchain_community==0.3.23
langchain_core==0.3.56
nltk==3.9.1
numpy==2.2.5
pandas==2.2.3
python-dotenv==1.1.0
rank_bm25==0.2.2
sentence_transformers==3.4.1
torch==2.5.1
tqdm==4.67.1
transformers==4.49.0
vllm==0.7.3
